+++ src/main/java/module-info.java
@@ -48,4 +48,6 @@ module pl.edu.pwr.pwrinspace.poliwrocket {
     opens pl.edu.pwr.pwrinspace.poliwrocket.Model.Protobuf to com.google.gson;
     exports pl.edu.pwr.pwrinspace.poliwrocket.Model.Command.Content;
     opens pl.edu.pwr.pwrinspace.poliwrocket.Model.Command.Content to com.google.gson;
+    exports pl.edu.pwr.pwrinspace.poliwrocket.Controller.Unused to javafx.base, javafx.controls, javafx.fxml, javafx.graphics, javafx.media, javafx.web, pl.edu.pwr.pwrinspace.poliwrocket.Controller;
+    opens pl.edu.pwr.pwrinspace.poliwrocket.Controller.Unused to javafx.base, javafx.controls, javafx.fxml, javafx.graphics, javafx.media, javafx.web;
 }
