+++ src/main/resources/Unused/TanwiarzView.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane prefHeight="335.0" prefWidth="410.0" styleClass="baseScene" stylesheets="@../Views/constants.css" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="pl.edu.pwr.pwrinspace.poliwrocket.Controller.TanwiarzController">
+<AnchorPane prefHeight="335.0" prefWidth="410.0" styleClass="baseScene" stylesheets="@../Views/constants.css" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="pl.edu.pwr.pwrinspace.poliwrocket.Controller.Unused.TanwiarzController">
    <children>
       <Tile fx:id="dataGauge1" layoutX="22.0" layoutY="27.0" maxHeight="90.0" maxWidth="-Infinity" prefHeight="90.0" prefWidth="170.0" skinType="SPARK_LINE" />
       <Tile fx:id="dataGauge2" averagingPeriod="100" layoutX="22.0" layoutY="127.0" maxHeight="90.0" maxWidth="-Infinity" prefHeight="90.0" prefWidth="170.0" skinType="SPARK_LINE" />
